work on tower placment and currency

+++ Assets/Scripts/Defenses/Defense.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Defense : MonoBehaviourPun {
 
 [SerializeField]string type = "Defense";
-[SerializeField]float cost = 0;
-public float Cost { get => cost; set => cost =  value ; }
+[SerializeField]int cost = 0;
+public int Cost { get => cost; set => cost = value ; }
 public string Type { get => type; set => type =  value ; }
 }
